fix(training): fix dtype casting issue
- pixel_values did not match the expected vae float16 type since
they were float32, casted to vae.dtype
- training, since DDP uses fp16, the loss was computed with same
which caused errors, now loss is calculate in float32 as expected

+++ training/loop.py
@@ -53,7 +53,7 @@ def encode_images_to_latents(pixel_values: torch.Tensor, vae) -> torch.Tensor:
         latents: (B, 4, 32, 32) tensor -- scaled latent representation
     """
 
-    latent_dist = vae.encode(pixel_values).latent_dist
+    latent_dist = vae.encode(pixel_values.to(dtype=vae.dtype)).latent_dist
     latents = latent_dist.sample()
     latents = latents * vae.config.scaling_factor
     return latents
@@ -154,7 +154,7 @@ def train_one_step(
     ).sample
 
     # 6. MSE loss between predicted and true noise
-    loss = F.mse_loss(noise_pred, noise)
+    loss = F.mse_loss(noise_pred.float(), noise.float())
 
     # 7. Backward pass (Accelerate handles gradient sync)
     accelerator.backward(loss)
